refactor(tests): more precise import names of components
