+++ src/components/TextField/TextField.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import { TextField, TextFieldProps } from '.';
+import { TextField, TextFieldProps } from './TextField';
 import '@testing-library/jest-dom';
 
 describe('TextField', () => {
